Added support for Garena client, and renamed Utils to LeagueMonitor for slight clarity
