+++ conduit/ProcessExtensions.cs
@@ -1,0 +1,26 @@
+﻿using System.Diagnostics;
+using System.Management;
+using System.Text;
+
+namespace MimicConduit
+{
+    public static class ProcessExtensions
+    {
+        public static string GetCommandLine(this Process process)
+        {
+            var commandLine = new StringBuilder(process.MainModule.FileName);
+
+            commandLine.Append(" ");
+            using (var searcher = new ManagementObjectSearcher("SELECT CommandLine FROM Win32_Process WHERE ProcessId = " + process.Id))
+            {
+                foreach (var @object in searcher.Get())
+                {
+                    commandLine.Append(@object["CommandLine"]);
+                    commandLine.Append(" ");
+                }
+            }
+
+            return commandLine.ToString();
+        }
+    }
+}
